Adapt the build system to create a 32-bit runtime on demand
It turns out that CMake doesn't support including a subdirectory multiple times
with different configurations. So, in order to build the runtime once as a
64-bit library and once for 32-bit targets, I define it as an external project,
which essentially causes CMake to invoke a completely separate process to build
it.

The clang wrappers now check whether "-m32" is set as a command-line argument
and link against the corresponding runtime.

The runtime doesn't support 128-bit integers in 32-bit mode, which is fine
because the compiler should never use them when compiling for a 32-bit
architecture.